RepoSack: New API to reload system repo
Add new API to drop all solvables from the system repo and load fresh
data from rpm database.
Re-creating the Base object is an expensive operation and there is a
use case in dnfdaemon to re-use the current session (which is basically
a wrapper arount the Base instance) even after executing an rpm
transaction - even for queries, or for running another transaction. With
this patch the system repo will contain correct information.